+++ src/main/kotlin/racing/Main.kt
@@ -1,5 +1,8 @@
 package racing
 
+import racing.domain.CarRacing
+import racing.ui.RacingInputReceiver
+
 fun main() {
     CarRacing().run(RacingInputReceiver.receive())
 }
