+++ src/main/kotlin/racing/domain/PowerStrategy.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 import kotlin.random.Random
