+++ src/main/kotlin/racing/ui/RacingInputReceiver.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.ui
 
 object RacingInputReceiver {
