+++ src/main/kotlin/racing/ui/UserInput.kt
@@ -1,3 +1,3 @@
-package racing
+package racing.ui
 
 data class UserInput(val carNames: List<String>, val moveCount: Int)
