+++ src/test/kotlin/racing/CarRacingTest.kt
@@ -3,6 +3,9 @@ package racing
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import racing.domain.CarRacing
+import racing.domain.StaticPowerStrategy
+import racing.ui.UserInput
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
