+++ src/test/kotlin/racing/RacingUITest.kt
@@ -8,7 +8,8 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.ArgumentsProvider
 import org.junit.jupiter.params.provider.ArgumentsSource
-import org.junit.jupiter.params.provider.CsvSource
+import racing.domain.Car
+import racing.ui.RacingUI
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
 import java.util.stream.Stream
